# openapi.yaml
openapi: 3.0.0
info:
  title: ShareTracker API
  version: 1.0.0
description: API documentation for the ShareTracker services.
servers:
  - url: http://localhost:5000
    description: Auth Service
  - url: http://localhost:5002
    description: Stock Service
paths:
  /register:
    post:
      summary: Register a new user
      description: Registers a new user with username, email, and password.
      tags:
        - Auth
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                username:
                  type: string
                email:
                  type: string
                password:
                  type: string
              required:
                - username
                - email
                - password
      responses:
        '201':
          description: User registered successfully
        '400':
          description: Missing required fields
        '409':
          description: Username or email already exists
  /login:
    post:
      summary: Login a user
      description: Authenticates a user and returns access and refresh tokens.
      tags:
        - Auth
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                username_or_email:
                  type: string
                password:
                  type: string
              required:
                - username_or_email
                - password
      responses:
        '200':
          description: Login successful
        '400':
          description: Missing required fields
        '401':
          description: Invalid credentials
  /quote/{symbol}:
    get:
      summary: Get stock quote
      description: Fetches stock quote data for a given symbol.
      tags:
        - Stock
      parameters:
        - name: symbol
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Stock quote data
        '401':
          description: Missing or invalid token
        '500':
          description: Failed to fetch quote
  /company/{symbol}:
    get:
      summary: Get company profile
      description: Fetches company profile data for a given symbol.
      tags:
        - Stock
      parameters:
        - name: symbol
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Company profile data
        '401':
          description: Missing or invalid token
        '500':
          description: Failed to fetch company profile
  /stock-candles:
    get:
      summary: Get stock candles (historical data)
      description: Fetches historical OHLC (Open, High, Low, Close) data for a given symbol using yfinance.
      tags:
        - Stock
      parameters:
        - name: symbol
          in: query
          required: true
          schema:
            type: string
        - name: period
          in: query
          required: false
          schema:
            type: string
            default: 1y
        - name: interval
          in: query
          required: false
          schema:
            type: string
            default: 1d
      responses:
        '200':
          description: Historical stock data
        '400':
          description: Missing symbol parameter
        '401':
          description: Missing or invalid token
        '404':
          description: No data found for this symbol
        '500':
          description: Failed to fetch candles